new feature: GET request only API server, log as as Admin to PUT, UPDATE, DELETE, or CREATE.

+++ src/pages/api/v1/shoes/[slug].ts
@@ -2,6 +2,7 @@ import { prisma } from "../../../../lib/prisma";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { stringifyTheDates } from "../../../../helper/stringify";
 import { ApiError, ShoePost, ShoeWithStringDates } from "../../../../interface";
+import { getSession } from "next-auth/react";
 
 type Data = {
   shoes: ShoeWithStringDates[];
@@ -12,6 +13,7 @@ export default async function handler(
   res: NextApiResponse<Data | ApiError>
 ) {
   const SLUG = req.query.slug.toString();
+
   if (req.method === "GET") {
     try {
       const shoe = await prisma.shoes.findUnique({ where: { slug: SLUG } });
@@ -29,41 +31,53 @@ export default async function handler(
     }
   }
 
-  if (req.method === "DELETE") {
-    try {
-      const shoe = await prisma.shoes.delete({ where: { slug: SLUG } });
-      if (shoe) {
-        const datesAsStrings = stringifyTheDates([
-          shoe,
-        ]) as ShoeWithStringDates[];
-        res.status(200).json({ shoes: datesAsStrings });
-      } else {
-        res.status(404).json({ error: `couldn't find rubber id: ${SLUG}` });
-      }
-    } catch (error) {
-      console.log(error);
-      res.status(500).json({ error: `server error` });
-    }
-  }
-  if (req.method === "PUT") {
-    try {
-      const shoeData: ShoePost = req.body;
+  const session = await getSession({ req });
 
-      const savedShoe = await prisma.shoes.update({
-        where: {
-          slug: SLUG,
-        },
-        data: shoeData,
-      });
-      if (savedShoe) {
-        const datesAsStrings = stringifyTheDates([
-          savedShoe,
-        ]) as ShoeWithStringDates[];
-        res.status(200).json({ shoes: datesAsStrings });
-      } else res.status(404).json({ error: `shoe not found` });
-    } catch (error) {
-      console.log(error);
-      res.status(500).json({ error: `server error` });
+  if (req.method !== "GET") {
+    if (session) {
+      if (req.method === "DELETE") {
+        try {
+          const shoe = await prisma.shoes.delete({ where: { slug: SLUG } });
+          if (shoe) {
+            const datesAsStrings = stringifyTheDates([
+              shoe,
+            ]) as ShoeWithStringDates[];
+            res.status(200).json({ shoes: datesAsStrings });
+          } else {
+            res.status(404).json({ error: `couldn't find rubber id: ${SLUG}` });
+          }
+        } catch (error) {
+          console.log(error);
+          res.status(500).json({ error: `server error` });
+        }
+      }
+      if (req.method === "PUT") {
+        try {
+          const shoeData: ShoePost = req.body;
+
+          const savedShoe = await prisma.shoes.update({
+            where: {
+              slug: SLUG,
+            },
+            data: shoeData,
+          });
+          if (savedShoe) {
+            const datesAsStrings = stringifyTheDates([
+              savedShoe,
+            ]) as ShoeWithStringDates[];
+            res.status(200).json({ shoes: datesAsStrings });
+          } else res.status(404).json({ error: `shoe not found` });
+        } catch (error) {
+          console.log(error);
+          res.status(500).json({ error: `server error` });
+        }
+      }
+    } else {
+      console.log(
+        "Must be logged in as ADMIN for anything other than a GET request"
+      );
+      res.status(401);
     }
+    res.end();
   }
 }
